% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComBat_met.r
\name{ComBat_met}
\alias{ComBat_met}
\title{Adjust for batch effects using a beta regression framework in DNA methylation data}
\usage{
ComBat_met(
  bv,
  batch,
  group = NULL,
  covar_mod = NULL,
  full_mod = TRUE,
  shrink = FALSE,
  mean.only = FALSE,
  feature.subset.n = NULL,
  pseudo_beta = 1e-04,
  ref.batch = NULL
)
}
\arguments{
\item{bv}{matrix of beta-values}

\item{batch}{vector for batch}

\item{group}{optional vector for biological condition of interest}

\item{covar_mod}{optional model matrix representing co-variates to be included in the model}

\item{full_mod}{Boolean variable indicating whether to include biological condition of interest in the model}

\item{shrink}{Boolean variable indicating whether to apply EB-shrinkage on parameter estimation}

\item{mean.only}{Boolean variable indicating whether to apply EB-shrinkage on the estimation of dispersion effects}

\item{feature.subset.n}{number of features to use in non-parametric EB estimation, only useful when shrink equals TRUE}

\item{pseudo_beta}{pseudo beta-values to be used for replacing extreme 0 and 1 beta-values. 
Value needs to be between 0 and 0.5.}

\item{ref.batch}{NULL by default. If given, that batch will be selected as a reference for batch correction.}
}
\value{
\code{ComBat_met} returns a feature x sample beta-value matrix, adjusted for batch effects.
}
\description{
ComBat_met is a model adapted from ComBat and ComBat-seq using beta regression,
which specifically targets DNA methylation data.
}
\examples{
# Generate a random beta-value matrix
bv_mat <- matrix(runif(n = 400, min = 0, max = 1), nrow = 50, ncol = 8)
batch <- c(rep(1, 4), rep(2, 4))
group <- rep(c(0, 1), 4)

# Adjust for batch effects including biological conditions
adj_bv_mat <- ComBat_met(bv_mat, batch = batch, group = group, full_mod = TRUE)
# Adjust for batch effects without including biological conditions
adj_bv_mat <- ComBat_met(bv_mat, batch = batch, group = group, full_mod = FALSE)

}
