% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstRankRMT.R
\name{EstRankRMT}
\alias{EstRankRMT}
\title{Estimate rank using RMT, taken from isva package}
\usage{
EstRankRMT(data.m, rmax = 50)
}
\arguments{
\item{data.m}{matrix, features by samples}

\item{rmax}{maximum rank}
}
\value{
rank
}
\description{
Estimate rank using RMT, taken from isva package
}
