 \name{svadat} \docType{data} \alias{svadat} \title{Simulated gene expression data affected by a group variable and an unobserved factor.} \description{ This data set is a simulated gene expression matrix with N(0,1) background noise and affected by two variables. The group variable affects the first 300 genes and the latent variable affects genes 200-500.  } \usage{svadat} \format{A 1000 x 20 gene expression matrix with genes in rows and arrays in columns} \keyword{datasets}