% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sva.check}
\alias{sva.check}
\title{A function for post-hoc checking of an sva object to check for
degenerate cases.}
\usage{
sva.check(svaobj, dat, mod, mod0)
}
\arguments{
\item{svaobj}{The transformed data matrix with the variables in rows and samples in columns}

\item{dat}{}

\item{mod}{The model matrix being used to fit the data}

\item{mod0}{The null model matrix being used to fit the data}
}
\value{
sv The estimated surrogate variables, one in each column

pprob.gam: A vector of the posterior probabilities each gene is affected by heterogeneity

pprob.b A vector of the posterior probabilities each gene is affected by mod

n.sv The number of significant surrogate variables
}
\description{
This function is designed to check for degenerate cases in the
sva fit and fix the sva object where possible.
}
\details{
\code{\link{empirical.controls}} for a direct estimate of the empirical controls.
}

